alert_system/requirements.txt
python-dotenv
pandas
watchdog
